Cash dev

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -29,7 +29,7 @@ public class CashMachine {
         );
     }
 
-    public void deposit(int amount) {
+    public void deposit(double amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -38,7 +38,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(double amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
@@ -72,7 +72,7 @@ public class CashMachine {
     }
 
     public String acctBalance(){
-        return Integer.toString(accountData.getBalance());
+        return Double.toString(accountData.getBalance());
     }
 
     private <T> void tryCall(Supplier<ActionResult<T> > action, Consumer<T> postAction) {
